move schedule hooks and types
