+++ apps/web/modules/schedules/hooks/useNonEmptyScheduleDays.ts
@@ -1,8 +1,8 @@
 import { useMemo } from "react";
 
-import type { Slots } from "../use-schedule/types";
+import type { Slots } from "~/schedules/lib/types";
 
-export const getNonEmptyScheduleDays = (slots?: Slots) => {
+const getNonEmptyScheduleDays = (slots?: Slots) => {
   if (typeof slots === "undefined") return [];
 
   const nonEmptyDays: string[] = [];
