+++ apps/web/modules/schedules/lib/types.ts
